ace-validators/requirements.txt
SQLAlchemy~=1.4.46
alembic
pymysql
passlib
fastapi==0.99.1
fastapi-scheduler==0.0.12
fastapi_amis_admin==0.5.7
uvicorn<=0.18.2
async-exit-stack
async-generator
requests
PyJWT
cryptography
slackclient
flake8
pytest
pytz
pyzmq
redis
schedule
python-dateutil
python-multipart
structlog
boto3
altonomy-settlement-engine
altonomy-client>=0.4.37
aiosqlite
slack-sdk
pandas
dataframe_image
lxml
tabulate
matplotlib
sqlmodel==0.0.11
